put an alert msg if there's a sub categories

+++ supprimer.php
@@ -4,27 +4,30 @@
     if (isset($_GET['id'])) {
     $id=$_GET['id'];
     if ($id != ""){
-        mysqli_query($con,"DELETE FROM achat_fornisseur WHERE ID_ARTICLE in (select ID_ARTICLE from article join categore2 join categore1 on categore2.ID_CATEGORE2=article.ID_CATEGORE2 and categore2.ID_CATEGORE1=categore1.ID_CATEGORE1
-        where categore1.ID_CATEGORE1=$id)");
-        mysqli_query($con,"DELETE FROM reception WHERE ID_ARTICLE in (select ID_ARTICLE from article join categore2 join categore1 on categore2.ID_CATEGORE2=article.ID_CATEGORE2 and categore2.ID_CATEGORE1=categore1.ID_CATEGORE1
-        where categore1.ID_CATEGORE1=$id)");
-        mysqli_query($con,"");
-        mysqli_query($con,"DELETE FROM article WHERE ID_CATEGORE2 in (select ID_CATEGORE2 from categore2 where ID_CATEGORE1=$id)");
-        mysqli_query($con,"DELETE FROM categore2 WHERE ID_CATEGORE1=$id");
-        mysqli_query($con,"DELETE FROM categore1 WHERE ID_CATEGORE1=$id");
+        $result = mysqli_query($con,"SELECT * FROM categore2 WHERE ID_CATEGORE1=$id");
+        $row=mysqli_fetch_array($result);
+        if ( $row[0] == null) {
+            mysqli_query($con,"DELETE FROM categore1 WHERE ID_CATEGORE1=$id");
+            echo "<div class='alert alert-success text-center form-signin' role='alert'>La categorie a ete supprimer avec succes</div>";
+        }
+        else {
+            echo "<div class='alert alert-warning text-center form-signin' role='alert'>vous doit supprimer les sous catégories avant de supprimer cette catégorie car ils ont relient</div>";
+        }   
     }
     mysqli_close($con); 
 }
     if (isset($_GET['id2'])) {
         $id2=$_GET['id2'];
         if ($id2 != ""){
-            mysqli_query($con,"DELETE FROM achat_fornisseur WHERE ID_ARTICLE in (select ID_ARTICLE from article join categore2 join categore1 on categore2.ID_CATEGORE2=article.ID_CATEGORE2 and categore2.ID_CATEGORE1=categore1.ID_CATEGORE1
-            where categore1.ID_CATEGORE1=$id2)");
-            mysqli_query($con,"DELETE FROM reception WHERE ID_ARTICLE in (select ID_ARTICLE from article join categore2 join categore1 on categore2.ID_CATEGORE2=article.ID_CATEGORE2 and categore2.ID_CATEGORE1=categore1.ID_CATEGORE1
-            where categore1.ID_CATEGORE1=$id2)");
-            mysqli_query($con,"");
-            mysqli_query($con,"DELETE FROM article WHERE ID_CATEGORE2 in (select ID_CATEGORE2 from categore2 where ID_CATEGORE1=$id2)");
-            mysqli_query($con,"DELETE FROM categore2 WHERE ID_CATEGORE2=$id2");
+            $result2 = mysqli_query($con,"SELECT * FROM article WHERE ID_CATEGORE2=$id2");
+            $row2=mysqli_fetch_array($result2);
+            if ( $row2[0] == null) {
+                mysqli_query($con,"DELETE FROM categore2 WHERE ID_CATEGORE2=$id2");
+                echo "<div class='alert alert-success text-center form-signin' role='alert'>La categorie a ete supprimer avec succes</div>";
+            }
+            else {
+                echo "<div class='alert alert-warning text-center form-signin' role='alert'>vous doit supprimer les sous catégories avant de supprimer cette catégorie car ils ont relient</div>";
+            }            
         }
         mysqli_close($con);
     }
